feat: make both bus envelopes Avro-encodable by deriving each one's schema from a declared registry of body types, with a union branch per body and our own struct walk that skips unexported fields and honours avro:"-" before descending, so the schema describes exactly what encoding/json already puts on the wire, and narrow the two sync cursors on the wire DTOs to int64 because Avro has no unsigned 64-bit type

+++ internal/app/consumer/event_history_id_update.go
@@ -7,7 +7,7 @@ import (
 )
 
 func (s *JobsService) HandleHistoryIDUpdate(ctx context.Context, e *models.JobEventHistoryIDUpdate) error {
-	if err := s.EmailHistoryIDRepository.Put(ctx, e.UserID, e.EmailID, e.HistoryID); err != nil {
+	if err := s.EmailHistoryIDRepository.Put(ctx, e.UserID, e.EmailID, uint64(e.HistoryID)); err != nil {
 		CaptureError(e.UserID, e.EmailID, err)
 		return err
 	}
